Moved and renamed PROBChange to PROBTafTrend
