+++ src/main/java/com/mivek/model/trend/PROBTafTrend.java
@@ -1,17 +1,17 @@
-package com.mivek.model;
+package com.mivek.model.trend;
 
 import com.mivek.enums.WeatherChangeType;
-import com.mivek.model.trend.AbstractTafTrend;
+import com.mivek.model.Validity;
 
 /**
  * @author mivek
  */
-public class PROBChange extends AbstractTafTrend<Validity> {
+public class PROBTafTrend extends AbstractTafTrend<Validity> {
 
     /**
      * Constructor.
      */
-    public PROBChange() {
+    public PROBTafTrend() {
         super(WeatherChangeType.PROB);
     }
