+++ src/main/java/com/mivek/parser/TAFParser.java
@@ -5,7 +5,6 @@ import java.time.LocalTime;
 import com.mivek.model.Airport;
 import com.mivek.model.BeginningValidity;
 import com.mivek.model.Cloud;
-import com.mivek.model.PROBChange;
 import com.mivek.model.TAF;
 import com.mivek.model.TemperatureDated;
 import com.mivek.model.Validity;
@@ -14,6 +13,7 @@ import com.mivek.model.WeatherCondition;
 import com.mivek.model.trend.AbstractTafTrend;
 import com.mivek.model.trend.BECMGTafTrend;
 import com.mivek.model.trend.FMTafTrend;
+import com.mivek.model.trend.PROBTafTrend;
 import com.mivek.model.trend.TEMPOTafTrend;
 import com.mivek.utils.Converter;
 import com.mivek.utils.Regex;
@@ -156,7 +156,7 @@ public final class TAFParser extends AbstractParser<TAF> {
                 }
                 taf.addFM(change);
             } else if (parts[0].startsWith(PROB)) {
-                PROBChange change = new PROBChange();
+                PROBTafTrend change = new PROBTafTrend();
                 iterChanges(0, parts, change);
                 taf.addProb(change);
             }
